Moved function to avoid unnecessary import
